_legacy/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pyyaml>=6.0
python-multipart
requests
# For better performance
uvloop; sys_platform != 'win32'
